rename cv workspace
