+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CandidateCvSectionItemTimeRange.cs
