+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvAbout.cs
