+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvContent.cs
