+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvSection.cs
