+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvSectionItem.cs
