+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvSectionItemDescription.cs
