+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvSectionItemSortNumber.cs
