+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/CvSectionName.cs
