+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/OrganisationName.cs
