+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/SectionItemLocation.cs
