+++ src/CareerBoostAI.Domain/Candidate/CvEntity/ValueObjects/SequenceIndex.cs
