+++ src/CareerBoostAI.Domain/Candidate/Factories/ICandidateFactory.cs
@@ -1,6 +1,6 @@
-﻿using CareerBoostAI.Domain.Candidate.ValueObjects;
+﻿using CareerBoostAI.Domain.Candidate.CvEntity;
+using CareerBoostAI.Domain.Candidate.ValueObjects;
 using CareerBoostAI.Domain.Common.ValueObjects;
-using CareerBoostAI.Domain.ValueObjects;
 
 namespace CareerBoostAI.Domain.Candidate.Factories;
 
@@ -13,6 +13,6 @@ public interface ICandidateFactory
     
     public CandidateAggregate Create(CandidateId id, FirstName firstName,
         LastName lastName, DateOfBirth dateOfBirth,
-        Email email, PhoneNumber phoneNumber, IEnumerable<Cv.Cv> cvs);
+        Email email, PhoneNumber phoneNumber, IEnumerable<CvEntity.Cv> cvs);
     
 }
